Renamed the images for the scripts GUI (will also update them at some point)
Fixed the generate favourites script, and added an info section and the option to use your games directories.
(also made it so if there is not game tbn it uses the XBMC cached one)

Added a modules check for the 480p Game Loader script as these modules are required for it to function.

+++ Mod Files/scripts/XBMC4Kids Extras/480p Game Loader/default.py
@@ -13,6 +13,9 @@ import os, sys
 import xbmcgui, xbmcaddon
 import zipfile
 
+#####	Modules required
+limpp	= xbmc.translatePath( "special://xbmc/scripts/.modules/script.module.limpp" )
+xbeinfo	= xbmc.translatePath( "special://xbmc/scripts/.modules/script.module.xbeinfo" )
 
 Addon = xbmcaddon.Addon('480p Game Loader')
 
@@ -28,23 +31,30 @@ getSetting = Addon.getSetting
 
 print '[SCRIPT][%s] version %s initialized!' % (__scriptname__, __version__)
 
+#####	Check for modules I need and error if not found.
+if os.path.isdir( limpp ):
+	if os.path.isdir( xbeinfo ):
+		#####	Extract the loaders.zip on first run.
+		if os.path.isfile( __path__ + "\\resources\\lib\\loaders.zip" ): 
+			xbmcgui.Dialog().ok( "Set-up","","I need to extract the loader xbe files before I can do anything.","This only happens once." )
+			with zipfile.ZipFile( __path__ + "\\resources\\lib\\loaders.zip" ) as zf:
+				zf.extractall( __path__ + "\\resources\\lib\\" )
+			os.remove( __path__ + "\\resources\\lib\\loaders.zip" )
+		#####	Main section.
+		if os.path.isdir( __path__ + "\\resources\\lib\\loaders" ): 
+			if (__name__ == "__main__"):
+				import resources.lib.__init__ as __init__
+				ui = __init__.GUI('%s.xml' % "main", __path__, 'default')
+				ui.doModal()
+				print '[SCRIPT][%s] version %s exited!' % (__scriptname__, __version__)
+				del ui
+		else:
+			xbmcgui.Dialog().ok( "Error","","The loaders folder is missing, please reinstall this script." )
 
-#####	extract the loaders.zip on first run.
-if os.path.isfile( __path__ + "\\resources\\lib\\loaders.zip" ): 
-	xbmcgui.Dialog().ok( "Set-up","","I need to extract the loader xbe files before I can do anything.","This only happens once." )
-	with zipfile.ZipFile( __path__ + "\\resources\\lib\\loaders.zip" ) as zf:
-		zf.extractall( __path__ + "\\resources\\lib\\" )
-	os.remove( __path__ + "\\resources\\lib\\loaders.zip" )
-	
-if os.path.isdir( __path__ + "\\resources\\lib\\loaders" ): 
-	if (__name__ == "__main__"):
-		import resources.lib.__init__ as __init__
-		ui = __init__.GUI('%s.xml' % "main", __path__, 'default')
-		ui.doModal()
-		print '[SCRIPT][%s] version %s exited!' % (__scriptname__, __version__)
-		del ui
+
+		sys.modules.clear()
+	else:
+		xbmcgui.Dialog().ok( "Error","","Module is missing.","[B]script.module.xbeinfo[/B]" )
 else:
-	xbmcgui.Dialog().ok( "Error","","The loaders folder is missing, please reinstall this script." )
-
-
-sys.modules.clear()
+	xbmcgui.Dialog().ok( "Error","","Module is missing.","[B]script.module.limpp[/B]" )
+
